passed logout

+++ server/src/main/java/service/UserService.java
@@ -5,6 +5,7 @@ import dataaccess.DataAccessException;
 import dataaccess.UserDAO;
 import model.AuthData;
 import model.UserData;
+import request.LogoutRequest;
 import result.LoginResult;
 
 public class UserService {
@@ -50,8 +51,11 @@ public class UserService {
     }
 
     // Log out a user by invalidating their auth token
-    public void logout(AuthData auth) throws DataAccessException {
+    public void logout(LogoutRequest auth) throws ServiceException, DataAccessException {
         // Invalidate the user's auth token by deleting it
+        if (authDAO.getAuth(auth.authToken()) == null) {
+            throw new ServiceException(401, "Error: unauthorized");
+        }
         authDAO.deleteAuth(auth.authToken());
     }
